refactor: extract dynamic plugin trust evaluation
